Checkpoint: Move everything to the sqs dir
